+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+  * Make audio and video RTCP report intervals configurable #1021
   * Stop padding media packets, no crypto need it #1023
   * Reset SCTP streams when closing data channels and safely reuse stream IDs #1010
   * Fix small padding invisible to the pacer #1020